Fix entities

+++ ejbModule/com/entities/Estudiante.java
@@ -18,7 +18,7 @@ public class Estudiante implements Serializable {
 	@SequenceGenerator(name="ESTUDIANTES_IDESTUDIANTE_GENERATOR", sequenceName="estud_id_seq", allocationSize=1, initialValue=1)
 	@GeneratedValue(strategy=GenerationType.SEQUENCE, generator="ESTUDIANTES_IDESTUDIANTE_GENERATOR")
 	@Column(name="ID_ESTUDIANTE", unique=true, nullable=false, precision=38)
-	private int idEstudiante;
+	private long idEstudiante;
 
 	@Column(name="GENERACION", nullable=false, length=4)
 	private String generacion;
@@ -40,7 +40,7 @@ public class Estudiante implements Serializable {
 		this.generacion = generacion;
 	}
 
-	public int getIdEstudiante() {
+	public long getIdEstudiante() {
 		return this.idEstudiante;
 	}
